Revamp setup signals autotrade to reduce number of args and add new half_spread based on MAs
